requirements.txt
django
djangorestframework
django-guardian
django-crispy-forms